switch sg_admin.cpp to glm
